Openbox: Update syntax to match upstream...
Unfortunately the nature of a large config file like rc.xml
is that as soon as you try to tweak it you've effectively
forked it.

I'm trying to keep my version as close to the original form
as possible.

This commit makes all the non-semantic changes necessary to
make my fork match the upstream.  Further commits will surely
follow.