Docker-deliver: Stop using named identifier when saving...
I only did that because it was a way of forcing the identifier
(i.e the tag) to be included in the tar file so that the tag would
be loaded into the destination.

But now I am explicitly tagging, and in the future I might
generate my own "repositories" file in the tar output.